Added .travis.yml
There are many mirrors of nettle on github. Many other github projects
using automatic build feature.

This enhancement will automatic check and test the changes done. It will
improve the testing capability of component as build can be tested at
every pull request before commit.

To enable this feature , you will have to login in
https://travis-ci.org/.  using github account and enable the travis
feature for the nettle library.

This feature is useful for github mirrors of nettle.

[DES: improved .travis.yml]

Signed-off-by: Dmitry Eremin-Solenikov <dbaryshkov@gmail.com>